use common method to print extensions
