+++ ca/django_ca/management/commands/view_ca.py
@@ -50,7 +50,7 @@ class Command(BaseCommand):
         self.stdout.write('')
         self.stdout.write('X509 v3 certificate extensions for CA:')
         for name, value in ca.extensions():
-            self.stdout.write('%s:\n    %s' % (name, value))
+            self.print_extensions(ca)
 
         self.stdout.write('')
         self.stdout.write('X509 v3 certificate extensions for signed certificates:')
